Refactor auth controller: implement admin registration, login, token refresh, and logout functionalities with proper error handling and response structure.
